--- src/main/java/interface_adapters/ChangeProfileOutputAdapter.java
@@ -1,28 +1,0 @@
-package interface_adapters;
-
-import usecases.profile_changes.ChangeProfileOutputBoundary;
-import usecases.profile_changes.ChangeProfileOutputModel;
-
-public class ChangeProfileOutputAdapter implements ChangeProfileOutputBoundary {
-
-
-    @Override
-    public void setPic(ChangeProfileOutputModel inputModel) {
-
-    }
-
-    @Override
-    public void delPic(ChangeProfileOutputModel inputModel) {
-
-    }
-
-    @Override
-    public void updateDescr(ChangeProfileOutputModel inputModel) {
-
-    }
-
-    @Override
-    public void updateName(ChangeProfileOutputModel inputModel) {
-
-    }
-}
