+++ src/main/java/interface_adapters/change_profile/ChangeProfileController.java
@@ -1,4 +1,4 @@
-package interface_adapters;
+package interface_adapters.change_profile;
 
 public class ChangeProfileController {
